6.10 Update Map Intent
