+++ app/src/main/res/menu/forecastfragment.xml
@@ -2,7 +2,10 @@
 
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/action_refresh"
-        android:title="@string/action_refresh"
+    <item android:id="@+id/action_map"
+        android:title="@string/action_map"
         app:showAsAction="never" />
+    <!--<item android:id="@+id/action_refresh"-->
+    <!--android:title="@string/action_refresh"-->
+    <!--app:showAsAction="never" />-->
 </menu>
